Move csv to libs
